requirements.txt
python-multipart
fastapi
uvicorn[standard]
numpy
faiss-cpu
pytorch